requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
chardet==4.0.0
constantly==15.1.0
crispy-bootstrap5==0.4
cryptography==3.4.8
daphne==3.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.5
django-allauth==0.45.0
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-filter==2.4.0
django-heroku==0.3.1
django-redis==5.0.0
djangorestframework==3.12.4
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itsdangerous==2.0.1
Markdown==3.3.4
msgpack==1.0.2
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.0
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
python-dateutil==2.8.2
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requests-cache==0.7.2
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.2
url-normalize==1.4.3
urllib3==1.26.6
whitenoise==5.3.0
zope.interface==5.4.0
